Remove resource controller 'get_resources' method
The 'get_resources' method (to retrieve a list of all resource
instances) ignores the API pagination and is thus limits the return data
to 100 objects (default limit value).

After f6dc346 this method is no longer used so it's safe to remove it
from the library until it can be refactored with pagination.