Return region specific PowerVS storage types
PowerIaas objects are bound to the power cloud instance region. This
'get_storage_types' call to the broker API is (hopefully) ephemeral, and
it's replacement should only return storage types available in the
instance region.

If we parse this with in the method now, then we can drop in the
replacement API later without change the return data structure.