Remove CreateImage os_type property validation
The PCloudImages OpenAPI spec includes an enum for the CreateImage
"osType" property. Once an image is created, this value is returned as
an attribute in Image ("specifications.imageType") and PVMInstance
("osType") objects, but is no longer enumerated. I'm removing this
validation to allow the possibility of importing other OS types without
updating the SDK.